Smol refactor, remove commented code, fix slf4j bridge
