+++ src/main/scala/io/pg/Main.scala
@@ -23,7 +23,6 @@ import io.pg.Prelude._
 import org.http4s.HttpApp
 import org.http4s.server.blaze.BlazeServerBuilder
 import org.http4s.server.middleware
-import org.slf4j.impl.StaticLoggerBinder
 
 object Main extends IOApp {
 
@@ -47,7 +46,7 @@ object Main extends IOApp {
     console // |+| file
   }
     .evalTap { logger =>
-      Sync[F].delay(StaticLoggerBinder.globalLogger.set(logger.mapK(Effect.toIOK).some))
+      Sync[F].delay(OdinInterop.globalLogger.set(logger.mapK(Effect.toIOK).some))
     }
 
   def mkServer[F[_]: Logger: ConcurrentEffect: Timer](http: HttpConfig, meta: MetaConfig, routes: HttpApp[F]) = {
