+++ src/main/scala/io/pg/OdinInterop.scala
@@ -1,4 +1,4 @@
-package org.slf4j.impl
+package io.pg
 
 import java.util.concurrent.atomic.AtomicReference
 
@@ -8,24 +8,18 @@ import cats.effect.IO
 import io.odin.Logger
 import io.odin.slf4j.OdinLoggerBinder
 
-class StaticLoggerBinder extends OdinLoggerBinder[IO] {
+class OdinInterop extends OdinLoggerBinder[IO] {
   implicit val F: Effect[IO] = IO.ioEffect
   implicit val clock: Clock[IO] = Clock.create[IO]
 
   val loggers: PartialFunction[String, Logger[IO]] = {
-    val theLogger: String => Option[Logger[IO]] = _ => StaticLoggerBinder.globalLogger.get()
+    val theLogger: String => Option[Logger[IO]] = _ => OdinInterop.globalLogger.get()
 
     theLogger.unlift
   }
 
 }
 
-object StaticLoggerBinder extends StaticLoggerBinder {
+object OdinInterop {
   val globalLogger = new AtomicReference[Option[Logger[IO]]](None)
-
-  //EC isn't used - only Clock is required
-  val REQUESTED_API_VERSION: String = "1.7"
-
-  def getSingleton: StaticLoggerBinder = this
-
 }
