+++ src/main/scala/org/slf4j/impl/StaticLoggerBinder.java
@@ -1,0 +1,10 @@
+package org.slf4j.impl;
+
+public class StaticLoggerBinder extends io.pg.OdinInterop {
+  public static String REQUESTED_API_VERSION = "1.7";
+
+  private static final StaticLoggerBinder _instance = new StaticLoggerBinder();
+  public static StaticLoggerBinder getSingleton() {
+    return _instance;
+  }
+}
